+++ src/bus/bus_obj.h
@@ -13,7 +13,7 @@ protected:
   uint16_t last_addr;
   uint16_t size_addr;
 
-  uint16_t frequency;
+  uint32_t frequency;
 
 public:
   std::string name;
@@ -22,8 +22,8 @@ public:
   uint16_t get_size();
   uint16_t get_init_addr();
   uint16_t get_last_addr();
-  void     set_frequency(uint16_t);
-  uint16_t get_frequency();
+  void     set_frequency(uint32_t);
+  uint32_t get_frequency();
   virtual uint8_t read(uint16_t) = 0;
   virtual void write(uint16_t, uint8_t) = 0;
   virtual void step(Bus_obj*) = 0;
